backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.66.3
chromadb==0.6.3
pypdf==5.4.0
python-multipart==0.0.20
